Part4 Finish task 4.21, fix oopsie

+++ part4/controllers/blogs.js
@@ -11,22 +11,6 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-    try {
-        const blog = new Blog({
-            title: request.body.title,
-            author: request.body.author,
-            url: request.body.url,
-            likes: request.body.likes,
-            user: request.body.user._id,
-        })
-
-        const savedBlog = await blog.save()
-		user.blogs = user.blogs.concat(savedBlog._id)
-		await user.save()
-        response.status(201).json(savedBlog)
-    } catch(exception) {
-        response.status(400).end()
-    }
 	try {
 		const decodedToken = jwt.verify(request.token, process.env.SECRET)
 		if (!decodedToken.id) {
@@ -45,6 +29,9 @@ blogsRouter.post('/', async (request, response) => {
 		)
 		const savedBlog = await blog.save()
 
+		user.blogs = user.blogs.concat(savedBlog._id)
+		await user.save()
+
 		response.status(201).json(savedBlog)
 	} catch(exception) {
 		response.status(400).json({ error: exception }).end()
@@ -52,8 +39,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+	try {
+		const decodedToken = jwt.verify(request.token, process.env.SECRET)
+		if (!decodedToken.id) {
+			return response.status(401).json({ error: 'token invalid' })
+		}
+
+		const user = await User.findById(decodedToken.id)
+		const blog = await Blog.findById(request.params.id)
+		if ( blog.user.toString() === user.id.toString() ) {
+			await Blog.findByIdAndRemove(request.params.id)
+		}
+		response.status(204).end()
+	} catch(exception) {
+		response.status(400).json({ error: exception }).end()
+	}
 })
 
 blogsRouter.put('/:id', async (request, response) => {
